fix bem usage and common modal component definition
