+++ web/libs/core/src/index.ts
@@ -5,6 +5,7 @@ export * from "./lib/utils/analytics";
 export * from "./lib/utils/urlJSON";
 export * from "./lib/utils/helpers";
 export * from "./lib/utils/string";
+export * from "./lib/utils/bem";
 export * from "./hooks/useAbortController";
 export * from "./lib/hooks/useCopyText";
 export * from "./hooks/usePageTitle";
