+++ web/libs/ui/src/components/Modal/ModalCloseButton.tsx
@@ -6,10 +6,9 @@ export const ModalCloseButton = () => {
   const modal = useModalControls();
   return (
     <Button
-      look="text"
-      size="compact"
+      look="string"
       className="!p-0 [&_svg]:!w-6 [&_svg]:!h-6"
-      icon={<IconClose />}
+      leading={<IconClose />}
       aria-label="Close modal"
       onClick={() => modal?.hide()}
     />
