--- web/libs/ui/src/components/Modal/ModalContext.ts
@@ -1,3 +1,0 @@
-import { BemWithSpecificContext } from "./bem";
-
-export const { Block, Elem } = BemWithSpecificContext();
