+++ web/libs/ui/src/components/Modal/ModalContext.tsx
@@ -1,0 +1,3 @@
+import { BemWithSpecificContext } from "@humansignal/core";
+
+export const { Block, Elem } = BemWithSpecificContext();
